add switch storage in services

+++ src/main/java/by/itacademy/jd2/votetask/dao/sql/GenresDaoSql.java
@@ -3,8 +3,9 @@ package by.itacademy.jd2.votetask.dao.sql;
 import by.itacademy.jd2.votetask.dao.api.IGenresDao;
 import by.itacademy.jd2.votetask.dto.GenreDTO;
 import by.itacademy.jd2.votetask.exceptions.DataAccessException;
-import by.itacademy.jd2.votetask.util.DataSource;
+import by.itacademy.jd2.votetask.util.DataSourceHolder;
 
+import javax.sql.DataSource;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -19,10 +20,12 @@ public class GenresDaoSql implements IGenresDao<GenreDTO> {
     private static final String DELETE_QUERY = "DELETE from data.genres where id=?;";
     private static final String EXIST_QUERY = "SELECT EXISTS (SELECT * FROM data.genres WHERE id = ?);";
     private static final String UPDATE_QUERY = "UPDATE data.genres SET name = ? WHERE id=?;";
+    private static final String CHECK_VOTES_FOR_GENRE = "SELECT EXISTS (SELECT * FROM data.vote_genre WHERE id_genre = ?);";
+    private final DataSource dataSource = DataSourceHolder.getDataSource();
 
     @Override
     public void create(GenreDTO genreDTO) {
-        try (Connection connection = DataSource.getConnection();
+        try (Connection connection = dataSource.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(CREATE_QUERY)) {
             preparedStatement.setString(1, genreDTO.getTitle());
             preparedStatement.executeUpdate();
@@ -33,7 +36,7 @@ public class GenresDaoSql implements IGenresDao<GenreDTO> {
 
     @Override
     public List<GenreDTO> readAll() {
-        try (Connection connection = DataSource.getConnection();
+        try (Connection connection = dataSource.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(READ_ALL_QUERY);
              ResultSet resultSet = preparedStatement.executeQuery()) {
             List<GenreDTO> entityList = new ArrayList<>();
@@ -49,10 +52,15 @@ public class GenresDaoSql implements IGenresDao<GenreDTO> {
 
     @Override
     public boolean delete(Long id) {
-        try (Connection connection = DataSource.getConnection();
+        boolean isVoted = checkVotesForGenre(id);
+        if(isVoted){
+            return false;
+        }
+        try (Connection connection = dataSource.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(DELETE_QUERY)) {
             preparedStatement.setLong(1, id);
-          return preparedStatement.execute();
+            int affectedRows = preparedStatement.executeUpdate();
+            return affectedRows != 0;
         } catch (SQLException e) {
             throw new DataAccessException("SQLException deleteById method :" + e);
         }
@@ -60,7 +68,7 @@ public class GenresDaoSql implements IGenresDao<GenreDTO> {
 
     @Override
     public void update(GenreDTO genreDTO) {
-        try (Connection connection = DataSource.getConnection();
+        try (Connection connection = dataSource.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(UPDATE_QUERY)) {
             Long id = genreDTO.getId();
             String title = genreDTO.getTitle();
@@ -74,8 +82,21 @@ public class GenresDaoSql implements IGenresDao<GenreDTO> {
 
     @Override
     public boolean exist(Long id) {
-        try (Connection connection = DataSource.getConnection();
+        try (Connection connection = dataSource.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(EXIST_QUERY)) {
+            preparedStatement.setLong(1, id);
+            ResultSet resultSet = preparedStatement.executeQuery();
+            resultSet.next();
+            return resultSet.getBoolean("exists");
+        } catch (SQLException e) {
+            throw new DataAccessException("SQLException deleteById method :" + e);
+        }
+    }
+
+
+    private boolean checkVotesForGenre(Long id) {
+        try (Connection connection = dataSource.getConnection();
+             PreparedStatement preparedStatement = connection.prepareStatement(CHECK_VOTES_FOR_GENRE)) {
             preparedStatement.setLong(1, id);
             ResultSet resultSet = preparedStatement.executeQuery();
             resultSet.next();
